api/requirements.txt
flask
kafka-python
pymysql
cryptography